ch: refactor selectChannelType
